Revert "ExtensionProvidedDependencyParser"
This reverts commit 690b7c44e46eff856e5ede39b5dbbc7b15f47146.
