+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/parsers/v1/ExtensionProvidedDependencyParser.java
@@ -1,0 +1,44 @@
+package com.sap.cloud.lm.sl.mta.parsers.v1;
+
+import static com.sap.cloud.lm.sl.mta.handlers.v1.Schemas.EXT_PROVIDED_DEPENDENCY;
+
+import java.util.Map;
+
+import com.sap.cloud.lm.sl.common.ParsingException;
+import com.sap.cloud.lm.sl.mta.model.v1.ExtensionProvidedDependency;
+import com.sap.cloud.lm.sl.mta.model.v1.ExtensionProvidedDependency.Builder;
+import com.sap.cloud.lm.sl.mta.parsers.ModelParser;
+import com.sap.cloud.lm.sl.mta.schema.MapElement;
+
+public class ExtensionProvidedDependencyParser extends ModelParser<ExtensionProvidedDependency> {
+
+    protected static final String PROCESSED_OBJECT_NAME = "MTA extension provided dependency";
+
+    public static final String NAME = "name";
+    public static final String PROPERTIES = "properties";
+
+    public ExtensionProvidedDependencyParser(Map<String, Object> source) {
+        this(EXT_PROVIDED_DEPENDENCY, source);
+    }
+
+    protected ExtensionProvidedDependencyParser(MapElement schema, Map<String, Object> source) {
+        super(PROCESSED_OBJECT_NAME, schema, source);
+    }
+
+    @Override
+    public ExtensionProvidedDependency parse() throws ParsingException {
+        Builder builder = new Builder();
+        builder.setName(getName());
+        builder.setProperties(getProperties());
+        return builder.build();
+    }
+
+    protected String getName() {
+        return getStringElement(NAME);
+    }
+
+    protected Map<String, Object> getProperties() {
+        return getMapElement(PROPERTIES);
+    }
+
+}
