+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/parsers/v2/ExtensionProvidedDependencyParser.java
@@ -6,30 +6,16 @@ import java.util.Map;
 
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionProvidedDependency;
 import com.sap.cloud.lm.sl.mta.model.v2.ExtensionProvidedDependency.Builder;
-import com.sap.cloud.lm.sl.mta.parsers.ModelParser;
 import com.sap.cloud.lm.sl.mta.schema.MapElement;
 
-public class ExtensionProvidedDependencyParser extends ModelParser<ExtensionProvidedDependency> {
-
-    protected static final String PROCESSED_OBJECT_NAME = "MTA extension provided dependency";
-
-    public static final String NAME = "name";
-    public static final String PROPERTIES = "properties";
+public class ExtensionProvidedDependencyParser extends com.sap.cloud.lm.sl.mta.parsers.v1.ExtensionProvidedDependencyParser {
 
     public ExtensionProvidedDependencyParser(Map<String, Object> source) {
-        this(EXT_PROVIDED_DEPENDENCY, source);
+        super(EXT_PROVIDED_DEPENDENCY, source);
     }
 
     protected ExtensionProvidedDependencyParser(MapElement schema, Map<String, Object> source) {
-        super(PROCESSED_OBJECT_NAME, schema, source);
-    }
-
-    protected String getName() {
-        return getStringElement(NAME);
-    }
-
-    protected Map<String, Object> getProperties() {
-        return getMapElement(PROPERTIES);
+        super(schema, source);
     }
 
     @Override
